+++ crates/holochain_state/src/source_chain.rs
@@ -14,7 +14,6 @@ use holo_hash::DhtOpHash;
 use holo_hash::DnaHash;
 use holo_hash::EntryHash;
 use holo_hash::HasHash;
-use holochain_chc::*;
 use holochain_keystore::MetaLairClient;
 use holochain_sqlite::rusqlite;
 use holochain_sqlite::rusqlite::params;
@@ -231,12 +230,15 @@ impl SourceChain {
         .await
     }
 
+    /// Flush scratch records to the authored and DHT databases.
+    ///
+    /// `storage_arcs` should reflect the current target arcs for authored ops integration.
+    /// Returns a tuple of (`flushed_actions`, `total_inserted_warrants`).
     #[async_recursion]
-    #[cfg_attr(feature = "instrument", tracing::instrument(skip(self, chc)))]
+    #[cfg_attr(feature = "instrument", tracing::instrument(skip(self)))]
     pub async fn flush(
         &self,
         storage_arcs: Vec<DhtArc>,
-        chc: Option<ChcImpl>,
     ) -> SourceChainResult<(Vec<SignedActionHashed>, u32)> {
         // Nothing to write
         if self.scratch.apply(|s| s.is_empty())? {
@@ -287,7 +289,7 @@ impl SourceChain {
         // before starting any database read/write operations.
         let write_permit = self.vault.acquire_write_permit().await?;
 
-        // If there are records to write, then we need to respect the chain lock and push to the CHC
+        // If there are records to write, then we need to respect the chain lock
         if !records.is_empty() {
             self.vault
                 .read_async({
@@ -323,34 +325,6 @@ impl SourceChain {
                     }
                 })
                 .await?;
-
-            // Sync with CHC, if CHC is present
-            if let Some(chc) = chc.clone() {
-                // Skip the CHC sync if this is a countersigning session.
-                // If the session times out, we might roll the chain back to the previous head, and so
-                // we don't want the record to exist remotely.
-                if is_countersigning_session {
-                    tracing::debug!(
-                        "Skipping CHC push for countersigning session: {:?}",
-                        records
-                    );
-                } else {
-                    let payload = AddRecordPayload::from_records(
-                        self.keystore.clone(),
-                        (*self.author).clone(),
-                        records,
-                    )
-                    .await
-                    .map_err(SourceChainError::other)?;
-
-                    match chc.add_records_request(payload).await {
-                        Err(e @ ChcError::InvalidChain(_, _)) => Err(
-                            SourceChainError::ChcHeadMoved("SourceChain::flush".into(), e),
-                        ),
-                        e => e.map_err(SourceChainError::other),
-                    }?;
-                }
-            }
         }
 
         let chain_flush_result = self
@@ -424,7 +398,7 @@ impl SourceChain {
                             scratch.add_entry(entry, ChainTopOrdering::Relaxed);
                         }
                     })?;
-                    child_chain.flush(storage_arcs, chc).await
+                    child_chain.flush(storage_arcs).await
                 } else {
                     Err(SourceChainError::HeadMoved(
                         actions,
@@ -1170,7 +1144,6 @@ pub async fn genesis(
     dna_hash: DnaHash,
     agent_pubkey: AgentPubKey,
     membrane_proof: Option<MembraneProof>,
-    chc: Option<ChcImpl>,
 ) -> SourceChainResult<()> {
     let dna_action = Action::Dna(Dna {
         author: agent_pubkey.clone(),
@@ -1219,23 +1192,6 @@ pub async fn genesis(
     let agent_entry = agent_entry.into_option();
 
     let mut ops_to_integrate = Vec::new();
-
-    if let Some(chc) = chc {
-        let payload = AddRecordPayload::from_records(
-            keystore.clone(),
-            agent_pubkey.clone(),
-            vec![dna_record, agent_validation_record, agent_record],
-        )
-        .await
-        .map_err(SourceChainError::other)?;
-
-        match chc.add_records_request(payload).await {
-            Err(e @ ChcError::InvalidChain(_, _)) => {
-                Err(SourceChainError::ChcHeadMoved("genesis".into(), e))
-            }
-            e => e.map_err(SourceChainError::other),
-        }?;
-    }
 
     let ops_to_integrate = authored
         .write_async(move |txn| {
@@ -1463,7 +1419,7 @@ mod tests {
             .await?;
 
         let storage_arcs = vec![DhtArc::Empty];
-        chain_1.flush(storage_arcs.clone(), None).await?;
+        chain_1.flush(storage_arcs.clone()).await?;
         let seq = db
             .write_async(move |txn| chain_head_db_nonempty(txn))
             .await?
@@ -1471,7 +1427,7 @@ mod tests {
         assert_eq!(seq, 3);
 
         assert!(matches!(
-            chain_2.flush(storage_arcs.clone(), None).await,
+            chain_2.flush(storage_arcs.clone()).await,
             Err(SourceChainError::HeadMoved(_, _, _, _))
         ));
         let seq = db
@@ -1480,7 +1436,7 @@ mod tests {
             .seq;
         assert_eq!(seq, 3);
 
-        chain_3.flush(storage_arcs, None).await?;
+        chain_3.flush(storage_arcs).await?;
         let seq = db
             .write_async(move |txn| chain_head_db_nonempty(txn))
             .await?
@@ -1543,7 +1499,7 @@ mod tests {
             .unwrap();
 
         let storage_arcs = vec![DhtArc::Empty];
-        chain_1.flush(storage_arcs.clone(), None).await?;
+        chain_1.flush(storage_arcs.clone()).await?;
         let seq = db
             .write_async(move |txn| chain_head_db_nonempty(txn))
             .await?
@@ -1551,11 +1507,11 @@ mod tests {
         assert_eq!(seq, 3);
 
         assert!(matches!(
-            chain_2.flush(storage_arcs.clone(), None).await,
+            chain_2.flush(storage_arcs.clone()).await,
             Err(SourceChainError::HeadMoved(_, _, _, _))
         ));
 
-        chain_3.flush(storage_arcs, None).await?;
+        chain_3.flush(storage_arcs).await?;
         let head = db
             .write_async(move |txn| chain_head_db_nonempty(txn))
             .await?;
@@ -1596,7 +1552,7 @@ mod tests {
 
         let result = chain.delete_valid_agent_pub_key().await;
         assert!(result.is_ok());
-        chain.flush(vec![DhtArc::Empty], None).await.unwrap();
+        chain.flush(vec![DhtArc::Empty]).await.unwrap();
 
         // Valid agent pub key has been deleted. Repeating the operation should fail now as no valid
         // pub key can be found.
@@ -1661,7 +1617,7 @@ mod tests {
                 .put_weightless(action_builder, Some(entry), ChainTopOrdering::default())
                 .await?;
 
-            chain.flush(storage_arcs.clone(), None).await.unwrap();
+            chain.flush(storage_arcs.clone()).await.unwrap();
             (action, entry_hash)
         };
 
@@ -1720,7 +1676,7 @@ mod tests {
             let action = chain
                 .put_weightless(action_builder, Some(entry), ChainTopOrdering::default())
                 .await?;
-            chain.flush(storage_arcs.clone(), None).await.unwrap();
+            chain.flush(storage_arcs.clone()).await.unwrap();
 
             (action, entry_hash)
         };
@@ -1784,7 +1740,6 @@ mod tests {
                 fake_dna_hash(1),
                 carol.clone(),
                 None,
-                None,
             )
             .await
             .unwrap();
@@ -1811,7 +1766,7 @@ mod tests {
             chain
                 .put_weightless(action_builder, None, ChainTopOrdering::default())
                 .await?;
-            chain.flush(storage_arcs.clone(), None).await.unwrap();
+            chain.flush(storage_arcs.clone()).await.unwrap();
         }
 
         // alice should get her author cap grant as always, independent of
@@ -1863,7 +1818,7 @@ mod tests {
             let action = chain
                 .put_weightless(action_builder, Some(entry), ChainTopOrdering::default())
                 .await?;
-            chain.flush(storage_arcs.clone(), None).await.unwrap();
+            chain.flush(storage_arcs.clone()).await.unwrap();
             (action, entry_hash)
         };
 
@@ -1898,7 +1853,6 @@ mod tests {
                     fake_dna_hash(1),
                     bob.clone(),
                     None,
-                    None,
                 )
                 .await
                 .unwrap();
@@ -1926,7 +1880,7 @@ mod tests {
             chain
                 .put_weightless(action_builder, None, ChainTopOrdering::default())
                 .await?;
-            chain.flush(storage_arcs.clone(), None).await.unwrap();
+            chain.flush(storage_arcs.clone()).await.unwrap();
         }
 
         // bob must not get unrestricted cap grant any longer
@@ -1992,7 +1946,7 @@ mod tests {
                 .put_weightless(action_builder, Some(entry), ChainTopOrdering::default())
                 .await?;
 
-            chain.flush(storage_arcs, None).await.unwrap();
+            chain.flush(storage_arcs).await.unwrap();
         }
 
         let actual_cap_grant = chain
@@ -2031,7 +1985,6 @@ mod tests {
             fixt!(DnaHash),
             (*author).clone(),
             None,
-            None,
         )
         .await
         .unwrap();
@@ -2062,7 +2015,7 @@ mod tests {
             .put_weightless(create, Some(entry), ChainTopOrdering::default())
             .await
             .unwrap();
-        source_chain.flush(vec![DhtArc::Empty], None).await.unwrap();
+        source_chain.flush(vec![DhtArc::Empty]).await.unwrap();
 
         vault
             .read_async({
@@ -2175,7 +2128,7 @@ mod tests {
                     scratch.add_entry(entry_hashed, ChainTopOrdering::Strict);
                 })
                 .unwrap();
-            chain.flush(vec![DhtArc::Empty], None).await.unwrap();
+            chain.flush(vec![DhtArc::Empty]).await.unwrap();
 
             action
         };
@@ -2335,7 +2288,7 @@ mod tests {
             .await;
         assert!(result.is_ok());
 
-        chain.flush(vec![DhtArc::Empty], None).await.unwrap();
+        chain.flush(vec![DhtArc::Empty]).await.unwrap();
 
         // zomes initialized should be true after init zomes has run
         let zomes_initialized = chain.zomes_initialized().await.unwrap();
@@ -2372,7 +2325,7 @@ mod tests {
             .unwrap();
 
         // Flush should write warrants to DHT database
-        let (actions, warrant_count) = chain.flush(vec![], None).await.unwrap();
+        let (actions, warrant_count) = chain.flush(vec![]).await.unwrap();
         assert!(actions.is_empty());
         assert_eq!(warrant_count, 1);
 
@@ -2408,7 +2361,7 @@ mod tests {
             .unwrap();
 
         // Flush should write warrant to DHT database
-        let (actions, warrant_count) = chain.flush(vec![], None).await.unwrap();
+        let (actions, warrant_count) = chain.flush(vec![]).await.unwrap();
         assert!(actions.is_empty());
         assert_eq!(warrant_count, 1);
 
@@ -2433,7 +2386,7 @@ mod tests {
             .unwrap();
 
         // Flush should not write duplicate warrant to DHT database
-        let (actions, warrant_count) = chain.flush(vec![], None).await.unwrap();
+        let (actions, warrant_count) = chain.flush(vec![]).await.unwrap();
         assert!(actions.is_empty());
         assert_eq!(warrant_count, 1); // rejected inserts are not reported by the insertion method, so this will indicate 1
 
@@ -2477,7 +2430,7 @@ mod tests {
             .unwrap();
 
         // Flush should not write warrant to DHT database
-        let (actions, warrant_count) = chain.flush(vec![], None).await.unwrap();
+        let (actions, warrant_count) = chain.flush(vec![]).await.unwrap();
         assert!(actions.is_empty());
         assert_eq!(warrant_count, 0);
 
@@ -2512,7 +2465,6 @@ mod tests {
                 keystore.clone(),
                 dna_hash,
                 agent_key.clone(),
-                None,
                 None,
             )
             .await